requirements.txt
fastapi
uvicorn[standard]
pillow
jinja2
python-dotenv